cat record start

+++ db/migrate/20190908143440_create_cats.rb
@@ -4,7 +4,7 @@ class CreateCats < ActiveRecord::Migration[5.1]
       t.string :cat_name
       t.binary :cat_image
       t.string :cat_sex
-      t.integer :cat_weight
+      t.float :cat_weight
       t.integer :cat_age
       t.references :user, foreign_key: true
